adapt rs_zktrie
