+++ Cargo.toml
@@ -9,19 +9,24 @@ links = "zktrie"
 # Companioned with zktrie module: v0.6
 
 [dependencies]
+zktrie_rust = { git = "https://github.com/scroll-tech/rust-zktrie", branch = "porting"}
 
 [build-dependencies]
 gobuild = { git = "https://github.com/scroll-tech/gobuild.git" }
 
-
 [dev-dependencies]
 hex = "0.4"
 halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2.git", tag = "v2022_09_10" }
-poseidon-circuit = { git = "https://github.com/scroll-tech/poseidon-circuit.git", branch = "sync-ff-0.13" }
+poseidon-circuit = { git = "https://github.com/scroll-tech/poseidon-circuit.git", branch = "scroll-dev-1201" }
 
 [patch."https://github.com/privacy-scaling-explorations/halo2.git"]
-halo2_proofs = { git = "https://github.com/scroll-tech/halo2.git", branch = "sync-ff-0.13" }
+halo2_proofs = { git = "https://github.com/scroll-tech/halo2.git", branch = "v1.1" }
+
+[patch."https://github.com/scroll-tech/rust-zktrie"]
+zktrie_rust = { path = "../rust-zktrie"}
+
 
 [features]
 dual_codehash = []
+rs_zktrie = []
 default = ["dual_codehash"]
