+++ rust-toolchain
@@ -1,1 +1,1 @@
-nightly-2022-12-10
+nightly-2023-12-03
