+++ src/constants.rs
@@ -1,0 +1,10 @@
+pub const HASHLEN: usize = 32;
+pub const FIELDSIZE: usize = 32;
+#[cfg(not(feature = "dual_codehash"))]
+pub const ACCOUNTFIELDS: usize = 4;
+#[cfg(feature = "dual_codehash")]
+pub const ACCOUNTFIELDS: usize = 5;
+pub const ACCOUNTSIZE: usize = FIELDSIZE * ACCOUNTFIELDS;
+pub type Hash = [u8; HASHLEN];
+pub type StoreData = [u8; FIELDSIZE];
+pub type AccountData = [[u8; FIELDSIZE]; ACCOUNTFIELDS];
